Some simplifications to TimescaleWriter
ListenWriter does not have any use of its own and only subclass is
TimescaleWriter. It only has a few utility methods which can be nicely
moved into TimescaleWriter.
